Gowtham: Committing changes on review comments for CAL-27 (Added Extensions class for Occasion Picker)
